data_eng/images/datasus/bronze_to_silver/requirements.txt
pandas
pyarrow
pysus
dbfread
google-cloud-storage
